Fixed audio playback of 5:1.
Former-commit-id: ae1d96c46c7ea992d4a673dabf892a2f4418f363

+++ mrViewer/src/core/mrvSwizzleAudio.h
@@ -8,11 +8,11 @@ template< typename T >
 struct Swizzle
 {
      T* ptr;
-     unsigned last;
+     size_t last;
      
      inline Swizzle(void* data, unsigned data_size ) :
      ptr( (T*) data ),
-     last( data_size / unsigned(8*sizeof(T)) )
+     last( data_size / 8*sizeof(T) )
      {
      }
